requirements.txt
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2